// projects/sysos-app-backups-manager/src/lib/types/mount-volume-snapshot.ts
import {
  ImConnection,
  ImDataObject,
  NetAppSnapshot,
  NetAppVolume,
  NetAppVserver,
  NetAppIface,
  VMWareHost
} from '@sysos/app-infrastructure-manager';

export interface MountVolumeSnapshot {
  storage: ImConnection;
  vserver: ImDataObject & { info: { data: NetAppVserver } };
  volume: ImDataObject & { info: { data: NetAppVolume } };
  snapshot: ImDataObject & { info: { data: NetAppSnapshot } };
  uuid: string;
  virtual: ImConnection;
  host: ImDataObject & { info: { data: VMWareHost } };
  iface: ImDataObject & { info: { data: NetAppIface } };
  volumeName?: string;
  datastorePath?: string;
}